fix(cli): cron utilities validate strictly and accept dow=7
Two latent issues in `getNextCronDelay` (used by `vercel dev`'s cron
simulator) and a small reorganisation:

- **Strict input validation.** `getNextCronDelay` runs
  `validateCronSchedule` upfront and returns `null` for syntactically
  valid but out-of-range expressions (`60 * * * *`, `0 24 * * *`, etc.)
  instead of silently producing a never-matching schedule.
- **`dow=7` as Sunday.** Day-of-week values of `7` are normalised onto
  `0`, matching `validateCronSchedule`. Previously `dow=7` schedules
  silently never fired in `vercel dev`.

`validateCronSchedule` (plus its `validateCronField` helper and
`CRON_FIELD_RANGES` table) is moved up from `commands/crons/add.ts` to
`util/cron.ts`. `add.ts` now imports and re-exports it so external
consumers keep working unchanged. The `CronFields` interface is also
exported now.
